ScanResult: Add list of scanned item
Use a RecyclerView to display the scans on the main activity. Once the ISBN scan succeeded, the result is added to the list of items to display

TODO: notify the view that the dataset changed, and update the display

Introduce "collection_id" data in advance, it will be used later.